moved delete profile function to a separate component
